MCR-2707 move class MCRObjectInfo to package datamodel.objectinfo
